SynthItemCommand refactored
Changed different bool values to a single boolVal in the command, to be
used by all boolean commands.

+++ audifier.cpp
@@ -23,7 +23,7 @@ float Audifier::process()
     {
         for(unsigned int i = 0; i < dataIndexes.size(); i++)
         {
-            sample += (dataItem->at(dataIndexes[i]) / 1000000000.0);
+            sample += (dataItem->at(static_cast<unsigned int>(dataIndexes[i])) / 1000000000.0);
         }
     }
